+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Scrollbar for episode list
+- Sleep timer audio fade-out and customization options
 
 ### Fixed
 
 - Episode list scroll lag
+- Sleep timer persists off-screen
 
 ## [v0.5.2] - 2026-07-05